requirements.txt
# Core frameworks
flask
streamlit
requests

# Model handling
transformers>=4.44.0
peft>=0.12.0
accelerate>=0.33.0
torch>=2.1.0
unsloth>=2024.5.10

# Image processing
pillow

# Optional for GPU acceleration
xformers
bitsandbytes

# Utility
numpy
